Incomplete: Adding adapted version of disk_space_usage_analysis scripts
